more refactoring of preview views
