+++ frontend/app/view/preview/preview-edit.tsx
@@ -1,0 +1,79 @@
+// Copyright 2025, Command Line Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import { tryReinjectKey } from "@/app/store/keymodel";
+import { CodeEditor } from "@/app/view/codeeditor/codeeditor";
+import { globalStore } from "@/store/global";
+import { adaptFromReactOrNativeKeyEvent, checkKeyPressed } from "@/util/keyutil";
+import { fireAndForget } from "@/util/util";
+import { Monaco } from "@monaco-editor/react";
+import { useAtomValue, useSetAtom } from "jotai";
+import type * as MonacoTypes from "monaco-editor/esm/vs/editor/editor.api";
+import { useEffect } from "react";
+import type { SpecializedViewProps } from "./preview";
+
+function CodeEditPreview({ model }: SpecializedViewProps) {
+    const fileContent = useAtomValue(model.fileContent);
+    const setNewFileContent = useSetAtom(model.newFileContent);
+    const fileInfo = useAtomValue(model.statFile);
+    const fileName = fileInfo?.name;
+    const blockMeta = useAtomValue(model.blockAtom)?.meta;
+
+    function codeEditKeyDownHandler(e: WaveKeyboardEvent): boolean {
+        if (checkKeyPressed(e, "Cmd:e")) {
+            fireAndForget(() => model.setEditMode(false));
+            return true;
+        }
+        if (checkKeyPressed(e, "Cmd:s") || checkKeyPressed(e, "Ctrl:s")) {
+            fireAndForget(model.handleFileSave.bind(model));
+            return true;
+        }
+        if (checkKeyPressed(e, "Cmd:r")) {
+            fireAndForget(model.handleFileRevert.bind(model));
+            return true;
+        }
+        return false;
+    }
+
+    useEffect(() => {
+        model.codeEditKeyDownHandler = codeEditKeyDownHandler;
+        return () => {
+            model.codeEditKeyDownHandler = null;
+            model.monacoRef.current = null;
+        };
+    }, []);
+
+    function onMount(editor: MonacoTypes.editor.IStandaloneCodeEditor, monaco: Monaco): () => void {
+        model.monacoRef.current = editor;
+
+        editor.onKeyDown((e: MonacoTypes.IKeyboardEvent) => {
+            const waveEvent = adaptFromReactOrNativeKeyEvent(e.browserEvent);
+            const handled = tryReinjectKey(waveEvent);
+            if (handled) {
+                e.stopPropagation();
+                e.preventDefault();
+            }
+        });
+
+        const isFocused = globalStore.get(model.nodeModel.isFocused);
+        if (isFocused) {
+            editor.focus();
+        }
+
+        return null;
+    }
+
+    return (
+        <CodeEditor
+            blockId={model.blockId}
+            text={fileContent}
+            filename={fileName}
+            fileinfo={fileInfo}
+            meta={blockMeta}
+            onChange={(text) => setNewFileContent(text)}
+            onMount={onMount}
+        />
+    );
+}
+
+export { CodeEditPreview };
